fix(runtime): stop naming libc::time_t for the session sweep bound
musl's `libc::time_t` is deprecated because libc will follow musl 1.2's
move to a 64-bit time_t (rust-lang/libc#1848), so every musl build of the
runtime warned at the terminal-session cleanup constant.

Type the constant as i32 and widen it into whatever `timespec::tv_sec` is
on the target (`.into()`), which is what the notice asks for: code that
does not depend on the alias. The value (10 seconds) and the deadline
arithmetic are unchanged on glibc, musl and Apple targets.

+++ crates/mangostudio-runtime/src/subprocess.rs
@@ -40,7 +40,9 @@ mod windows_job;
 /// A PTY `close` resolves only after that sweep, so tests bound `close` by this plus slack.
 #[cfg(all(unix, test))]
 pub(crate) const TERMINAL_SESSION_CLEANUP_BOUND: std::time::Duration =
-    std::time::Duration::from_secs(unix_guardian::TERMINAL_SESSION_CLEANUP_SECONDS.unsigned_abs());
+    std::time::Duration::from_secs(
+        unix_guardian::TERMINAL_SESSION_CLEANUP_SECONDS.unsigned_abs() as u64,
+    );
 
 /// PTY child with the same guardian or Job ownership as a bounded process.
 pub(crate) enum PtyChild {
